@@ -1,7 +1,7 @@
 package org.bilan.co.config;
 
-import org.dozer.DozerBeanMapper;
-import org.dozer.Mapper;
+import com.github.dozermapper.core.DozerBeanMapperBuilder;
+import com.github.dozermapper.core.Mapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -11,7 +11,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 public class GeneralConfig {
     @Bean
     public Mapper mapper() {
-        return new DozerBeanMapper();
+        return DozerBeanMapperBuilder.buildDefault();
     }
 
     @Bean
