@@ -1,11 +1,16 @@
 package org.bilan.co.domain.dtos.common;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
-import org.bilan.co.domain.dtos.forums.PostDto;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class PagedResponse<T> {
     private List<T> data;
     private Integer nPages;
